requirements.txt
google-generativeai>=0.7.0
google-adk>=0.1.6
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# Optional (if you want OR-Tools later)
ortools>=9.9.3963

# Optional: debugging + testing
rich>=13.0.0
pytest>=7.0.0
